requirements.txt
python-telegram-bot==21.2
requests==2.28.1
Pillow==9.2.0
sounddevice
scipy
httpx
netifaces